refactor: delete download manager
